+++ Activities/Cryptography/UiPath.Cryptography.Activities/NetCore/ViewModels/DecryptFileViewModel.cs
@@ -1,233 +1,89 @@
-﻿using System;
+using System;
 using System.Activities;
 using System.Activities.DesignViewModels;
 using System.Activities.ViewModels;
-using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
-using System.Security;
-using UiPath.Cryptography.Activities.Helpers;
 using UiPath.Cryptography.Enums;
 using UiPath.Platform.ResourceHandling;
-
-#pragma warning disable CS0618 // obsolete encryption algorithm
 
 namespace UiPath.Cryptography.Activities.NetCore.ViewModels
 {
     [ExcludeFromCodeCoverage]
-    public class DecryptFileViewModel : DesignPropertiesViewModel
+    public class DecryptFileViewModel : DecryptCryptoViewModelBase
     {
-        private readonly DataSource<string> _encodingDataSource;
         private InArgument<IResource> _backupInputFile;
         private InArgument<string> _backupInputFilePath;
 
-        /// <summary>
-        /// Basic constructor
-        /// </summary>
-        /// <param name="services"></param>
         public DecryptFileViewModel(IDesignServices services) : base(services)
         {
-            _encodingDataSource = EncodingHelpers.ConfigureEncodingDataSource();
         }
 
-        /// <summary>
-        /// A drop-down which enables you to select the decryption algorithm you want to use.
-        /// </summary>
-        public DesignProperty<EncryptionAlgorithm> Algorithm { get; set; } = new DesignProperty<EncryptionAlgorithm>();
-
-        /// <summary>
-        /// The key that you want to use to decrypt the specified file.
-        /// </summary>
-        public DesignInArgument<string> Key { get; set; } = new DesignInArgument<string>();
-
-        /// <summary>
-        /// The secure string used to decrypt the input file.
-        /// </summary>
-        public DesignInArgument<SecureString> KeySecureString { get; set; } = new DesignInArgument<SecureString>();
-
-        /// <summary>
-        /// Switches Key as string or secure string
-        /// </summary>
-        public DesignProperty<KeyInputMode> KeyInputModeSwitch { get; set; } = new DesignProperty<KeyInputMode>();
-
-        /// <summary>
-        /// A drop-down which enables you to select the encoding option you want to use.
-        /// </summary>
-        public DesignInArgument<string> KeyEncodingString { get; set; } = new() { Name = nameof(KeyEncodingString) };
-
-        /// <summary>
-        /// If a file already exists at the path specified in the Output path field, selecting this check box overwrites it.
-        /// </summary>
-        public DesignProperty<bool> Overwrite { get; set; } = new DesignProperty<bool>();
-
-        /// <summary>
-        /// Specifies if the automation should continue even when the activity throws an error.
-        /// </summary>
-        public DesignInArgument<bool> ContinueOnError { get; set; } = new DesignInArgument<bool>();
-
-        /// <summary>
-        /// The file that you want to decrypt.
-        /// </summary>
         public DesignInArgument<IResource> InputFile { get; set; } = new DesignInArgument<IResource>();
-
-        /// <summary>
-        /// The path to the file that you want to decrypt.
-        /// </summary>
         public DesignInArgument<string> InputFilePath { get; set; } = new DesignInArgument<string>();
-
-        /// <summary>
-        /// Switches File as IResource or string
-        /// </summary>
         public DesignProperty<FileInputMode> FileInputModeSwitch { get; set; } = new DesignProperty<FileInputMode>();
-
-        /// <summary>
-        /// The output path to the file that you want to decrypt.
-        /// </summary>
         public DesignInArgument<string> OutputFilePath { get; set; } = new DesignInArgument<string>();
-
-        /// <summary>
-        /// The file that you want to decrypt.
-        /// </summary>
+        public DesignProperty<bool> Overwrite { get; set; } = new DesignProperty<bool>();
         public DesignOutArgument<ILocalResource> DecryptedFile { get; set; } = new DesignOutArgument<ILocalResource>();
-
-        public DesignInArgument<string> PrivateKeyFilePath { get; set; } = new DesignInArgument<string>();
-        public DesignInArgument<System.Security.SecureString> Passphrase { get; set; } = new DesignInArgument<System.Security.SecureString>();
-        public DesignProperty<bool> VerifySignature { get; set; } = new DesignProperty<bool>();
-        public DesignInArgument<string> PublicKeyFilePath { get; set; } = new DesignInArgument<string>();
 
         protected override void InitializeModel()
         {
             base.InitializeModel();
-            var propertyOrderIndex = 1;
+            var orderIndex = 1;
 
             InputFile.IsPrincipal = true;
             InputFile.IsVisible = false;
-            InputFile.OrderIndex = propertyOrderIndex++;
+            InputFile.OrderIndex = orderIndex++;
 
             InputFilePath.IsPrincipal = true;
             InputFilePath.IsVisible = true;
-            InputFilePath.OrderIndex = propertyOrderIndex++;
+            InputFilePath.OrderIndex = orderIndex++;
 
             FileInputModeSwitch.IsVisible = false;
 
-            Algorithm.IsPrincipal = true;
-            Algorithm.OrderIndex = propertyOrderIndex++;
-            Algorithm.DataSource = DataSourceHelper.ForEnum(EncryptionAlgorithm.AES, EncryptionAlgorithm.AESGCM, EncryptionAlgorithm.DES, EncryptionAlgorithm.RC2, EncryptionAlgorithm.Rijndael, EncryptionAlgorithm.TripleDES, EncryptionAlgorithm.PGP);
-            Algorithm.Widget = new DefaultWidget { Type = ViewModelWidgetType.Dropdown };
-
-            Key.IsPrincipal = true;
-            Key.IsVisible = true;
-            Key.OrderIndex = propertyOrderIndex++;
-
-            KeySecureString.IsPrincipal = true;
-            KeySecureString.IsVisible = false;
-            KeySecureString.OrderIndex = propertyOrderIndex++;
-
-            KeyInputModeSwitch.IsVisible = false;
-
-            KeyEncodingString.IsPrincipal = false;
-            KeyEncodingString.IsVisible = true;
-            KeyEncodingString.OrderIndex = propertyOrderIndex++;
-
-            KeyEncodingString.DataSource = _encodingDataSource;
-            KeyEncodingString.Widget = new DefaultWidget { Type = ViewModelWidgetType.Dropdown, Metadata = new Dictionary<string, string>() };
-
-            _encodingDataSource.Data = EncodingHelpers.GetAvailableEncodings();
+            ConfigureAlgorithmAndKeyProperties(ref orderIndex);
 
             OutputFilePath.IsPrincipal = false;
             OutputFilePath.IsVisible = true;
             OutputFilePath.IsRequired = false;
-            OutputFilePath.OrderIndex = propertyOrderIndex++;
+            OutputFilePath.OrderIndex = orderIndex++;
 
             Overwrite.IsPrincipal = false;
-            Overwrite.OrderIndex = propertyOrderIndex++;
+            Overwrite.OrderIndex = orderIndex++;
             Overwrite.Widget = new DefaultWidget { Type = ViewModelWidgetType.Toggle };
 
-            ContinueOnError.IsPrincipal = false;
-            ContinueOnError.OrderIndex = propertyOrderIndex++;
-            ContinueOnError.Widget = new DefaultWidget { Type = ViewModelWidgetType.NullableBoolean };
-            ContinueOnError.Value = false;
-
-            MenuActionsBuilder<KeyInputMode>.WithValueProperty(KeyInputModeSwitch)
-                .AddMenuProperty(Key, KeyInputMode.Key)
-                .AddMenuProperty(KeySecureString, KeyInputMode.SecureKey)
-                .BuildAndInsertMenuActions();
+            ConfigureTailProperties(ref orderIndex);
+            ConfigureKeyInputModeMenuActions();
 
             MenuActionsBuilder<FileInputMode>.WithValueProperty(FileInputModeSwitch)
                 .AddMenuProperty(InputFile, FileInputMode.File)
                 .AddMenuProperty(InputFilePath, FileInputMode.FilePath)
                 .BuildAndInsertMenuActions();
 
-            DecryptedFile.OrderIndex = propertyOrderIndex++;
-
-            PrivateKeyFilePath.IsPrincipal = false;
-            PrivateKeyFilePath.IsVisible = false;
-            PrivateKeyFilePath.OrderIndex = propertyOrderIndex++;
-
-            Passphrase.IsPrincipal = false;
-            Passphrase.IsVisible = false;
-            Passphrase.OrderIndex = propertyOrderIndex++;
-
-            VerifySignature.IsPrincipal = false;
-            VerifySignature.IsVisible = false;
-            VerifySignature.OrderIndex = propertyOrderIndex++;
-            VerifySignature.Widget = new DefaultWidget { Type = ViewModelWidgetType.Toggle };
-
-            PublicKeyFilePath.IsPrincipal = false;
-            PublicKeyFilePath.IsVisible = false;
-            PublicKeyFilePath.OrderIndex = propertyOrderIndex;
+            DecryptedFile.OrderIndex = orderIndex++;
 
             _backupInputFile = InputFile.Value;
             _backupInputFilePath = InputFilePath.Value;
         }
 
-        /// <inheritdoc/>
         protected override void InitializeRules()
         {
             base.InitializeRules();
-            Rule(nameof(KeyInputModeSwitch), KeyInputModeChanged_Action);
             Rule(nameof(FileInputModeSwitch), FileInputModeChanged_Action);
-            Rule(nameof(Algorithm), AlgorithmChanged_Action);
-            Rule(nameof(VerifySignature), VerifySignatureChanged_Action);
         }
 
-        /// <inheritdoc/>
         protected override void ManualRegisterDependencies()
         {
             base.ManualRegisterDependencies();
-            RegisterDependency(KeyInputModeSwitch, nameof(KeyInputModeSwitch.Value), nameof(KeyInputModeSwitch));
             RegisterDependency(FileInputModeSwitch, nameof(FileInputModeSwitch.Value), nameof(FileInputModeSwitch));
-            RegisterDependency(Algorithm, nameof(Algorithm.Value), nameof(Algorithm));
-            RegisterDependency(VerifySignature, nameof(VerifySignature.Value), nameof(VerifySignature));
         }
 
-        /// <summary>
-        /// Key input Mode has changed. Set controls visibility based on selection
-        /// </summary>
-        private void KeyInputModeChanged_Action()
-        {
-            if (Algorithm.Value == EncryptionAlgorithm.PGP) return;
-            ResetAllKeyInputMode();
-            switch (KeyInputModeSwitch.Value)
-            {
-                case KeyInputMode.Key:
-                    Key.IsRequired = true;
-                    Key.IsVisible = true;
-                    break;
-                case KeyInputMode.SecureKey:
-                    KeySecureString.IsVisible = true;
-                    KeySecureString.IsRequired = true;
-                    break;
-                default:
-                    throw new NotImplementedException();
-            }
-        }
-
-        /// <summary>
-        /// File input Mode has changed. Set controls visibility based on selection
-        /// </summary>
         private void FileInputModeChanged_Action()
         {
-            ResetAllInputFile();
+            InputFile.IsRequired = false;
+            InputFile.IsVisible = false;
+            InputFilePath.IsRequired = false;
+            InputFilePath.IsVisible = false;
+
             switch (FileInputModeSwitch.Value)
             {
                 case FileInputMode.File:
@@ -253,53 +109,6 @@ namespace UiPath.Cryptography.Activities.NetCore.ViewModels
                 default:
                     throw new NotImplementedException();
             }
-        }
-
-        private void AlgorithmChanged_Action()
-        {
-            bool isPgp = Algorithm.Value == EncryptionAlgorithm.PGP;
-
-            // Symmetric properties: hide when PGP
-            Key.IsVisible = !isPgp && KeyInputModeSwitch.Value == KeyInputMode.Key;
-            KeySecureString.IsVisible = !isPgp && KeyInputModeSwitch.Value == KeyInputMode.SecureKey;
-            KeyEncodingString.IsVisible = !isPgp;
-
-            // PGP properties: show when PGP
-            PrivateKeyFilePath.IsVisible = isPgp;
-            PrivateKeyFilePath.IsRequired = isPgp;
-            Passphrase.IsVisible = isPgp;
-            Passphrase.IsRequired = isPgp;
-            VerifySignature.IsVisible = isPgp;
-            PublicKeyFilePath.IsVisible = isPgp && VerifySignature.Value;
-            PublicKeyFilePath.IsRequired = isPgp && VerifySignature.Value;
-
-            if (!isPgp)
-            {
-                KeyInputModeChanged_Action();
-            }
-        }
-
-        private void VerifySignatureChanged_Action()
-        {
-            if (Algorithm.Value != EncryptionAlgorithm.PGP) return;
-            PublicKeyFilePath.IsVisible = VerifySignature.Value;
-            PublicKeyFilePath.IsRequired = VerifySignature.Value;
-        }
-
-        private void ResetAllKeyInputMode()
-        {
-            Key.IsRequired = false;
-            Key.IsVisible = false;
-            KeySecureString.IsRequired = false;
-            KeySecureString.IsVisible = false;
-        }
-
-        private void ResetAllInputFile()
-        {
-            InputFile.IsRequired = false;
-            InputFile.IsVisible = false;
-            InputFilePath.IsRequired = false;
-            InputFilePath.IsVisible = false;
         }
     }
 }
